+++ MelonLoader/Melons/MelonHandler.cs
@@ -243,19 +243,30 @@ namespace MelonLoader
         internal static void OnFixedUpdate() => InvokeMelonMethod(ref _Mods, x => x.OnFixedUpdate());
         internal static void OnLateUpdate() => InvokeMelonMethod(x => x.OnLateUpdate());
         internal static void OnGUI() => InvokeMelonMethod(x => x.OnGUI());
-        internal static void OnPreferencesSaved(string filepath) => InvokeMelonMethod(x =>
+        internal static void BONEWORKS_OnLoadingScreen() => InvokeMelonMethod(ref _Mods, x => x.BONEWORKS_OnLoadingScreen());
+
+        internal static void OnPreferencesSaved() => InvokeMelonMethod(x =>
         {
             x.OnModSettingsApplied();
             x.OnPreferencesSaved();
+        });
+        internal static void OnPreferencesSaved(string filepath) => InvokeMelonMethod(x =>
+        {
+            x.OnModSettingsApplied();
             x.OnPreferencesSaved(filepath);
         });
-        internal static void OnPreferencesLoaded(string filepath) => InvokeMelonMethod(x =>
+
+        internal static void OnPreferencesLoaded() => InvokeMelonMethod(x =>
         {
             x.OnModSettingsApplied();
             x.OnPreferencesLoaded();
+        });
+        internal static void OnPreferencesLoaded(string filepath = null) => InvokeMelonMethod(x =>
+        {
+            x.OnModSettingsApplied();
             x.OnPreferencesLoaded(filepath);
         });
-        internal static void BONEWORKS_OnLoadingScreen() => InvokeMelonMethod(ref _Mods, x => x.BONEWORKS_OnLoadingScreen());
+
 
         private static bool SceneWasJustLoaded = false;
         private static int CurrentSceneBuildIndex = -1;