+++ MelonLoader/Preferences/MelonPreferences_Category.cs
@@ -171,6 +171,7 @@ namespace MelonLoader
             }
             if (printmsg)
                 MelonLogger.Msg($"MelonPreferences Saved to {currentfile.FilePath}");
+            MelonHandler.OnPreferencesSaved();
             MelonHandler.OnPreferencesSaved(currentfile.FilePath);
         }