+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 27. Added Old Category Purging to Start Screen Customization.
 28. Fixed Issue with Custom Start Screen Image Scanning not taking into account File Extension.
 29. Added Optional Random Theme Selection to Start Screen.
+30. Fixed Issue with OnPreferencesSaved and OnPreferencesLoaded being called Incorrectly.
 
 ---